RAW without interleaver, typesafety, API changes
This was originally meant only to eradicate the interleaving
code, so that the raw API could be slightly faster.

Turns out a lot had to be done for that, and the API changed
to reflect those changes.

The RAW_Encoder template now has enabled/disabled methods,
so that you can only use the right one, depending on
whether you choose interleaving or not.

While doing all of this, I realized a nice way to let the user
choose the correct symbol number for the block. +1 for typesafety.

And since the commit became really big, I threw in a couple
of useless style changes, bringing back everything to 80 columns.

Really, I should be doing smaller commits.

Signed-off-by: Luca Fulchir <luker@fenrirproject.org>