Fix issue #21 - Decoding fails to often
Loop from 'i' to 'M' instead of 'i' to 'L' during Gaussian elimination in the
second phase.

Signed-off-by: Anders Martinsson <anders.martinsson@intinor.se>

This probably comes from a ill-fated idea to optimize and stop after the
triangular matrix is parsed, but that prevented us from parsing
other repair symbols. Which is really wrong.

Good work for ticket #21

Signed-off-by: Luca Fulchir <luker@fenrirproject.org>