Add option to ill with zeros the lost symbols
Fixes #24

Sometimes it makes sense to have incomplete data.
This way we both de-interleave the incomplete data
and we give a std::vector<bool> of which bytes were
transmitted correctly and which were zeroed out.

Accessible through the end_of_input() calls, as it is the
only call that makes sense for this, since we need to
also ignore any other repair symbol.

Signed-off-by: Luca Fulchir <luker@fenrirproject.org>